ML Powered App/requirements.txt
numpy==1.26.4
scikit-learn==1.4.1.post1
pandas==2.2.1
seaborn==0.13.2
streamlit==1.29.0
fastapi==0.103.0
uvicorn==0.20.0
watchdog==2.1.6
psycopg2==2.9.9
python-dotenv==0.21.0
asyncpg==0.29.0
apache-airflow==2.8.3